requirements.txt
pytorch-lightning==1.6.0
torchvision==0.12.0
transformers==4.17.0
datasets==2.0.0
scikit-learn==1.0.2
